feature: if a user not is premium, when trying to upload task the modal is displayed
separated google and premium Buttons in layouts and implement in settings views and update logic
