+++ lib/src/widgets/ButtonUpload/ButtonUpload.dart
@@ -2,11 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:tasks_ia_koderx/src/widgets/Button/Button.dart';
 
 class Buttonupload extends StatelessWidget {
-  const Buttonupload({super.key});
+  const Buttonupload({super.key, this.click});
+  final VoidCallback? click;
 
   @override
   Widget build(BuildContext context) {
     return Button(
+      click: click,
       padding: EdgeInsets.all(0),
       width: 40,
       style: ButtonStyle(
